# golang_backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Ufaelectrto market",
        "title": "Ufaelectrto API",
        "contact": {
            "name": "API Creator",
            "url": "https://t.me/amirich18",
            "email": "adamstradvers@gmail.com"
        },
        "version": "1.3"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/order": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all user orders",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Get user orders",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Orders limit in response",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Response page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.CountResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "count": {
                                            "$ref": "#/definitions/dto.Count"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.GetOrder"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create order user order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Create order",
                "parameters": [
                    {
                        "description": "order info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.DefaultData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ReturnOrder"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/order/cancel": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Cancel user order",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Cancel order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order id",
                        "name": "order_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/order/payment_methods": {
            "get": {
                "description": "Get all payments methods",
                "tags": [
                    "Order"
                ],
                "summary": "Get payments methods",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.DefaultData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.PaymentMethod"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/order/pick_up_points": {
            "get": {
                "description": "Get all pick up points",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Get pick up points",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.DefaultData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.PickUpPoint"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/order/receipt": {
            "post": {
                "description": "Get order receipt",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Get receipt",
                "parameters": [
                    {
                        "type": "string",
                        "description": "promo code",
                        "name": "promo_code",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.DefaultData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.Receipt"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/order_products": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all user products",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Get user products",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Orders limit in response",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Response page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.CountResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "count": {
                                            "$ref": "#/definitions/dto.Count"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.UserOrdersProducts"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.CDEKDelivery": {
            "type": "object",
            "properties": {
                "pick_up_point_address": {
                    "description": "TODO continue logic",
                    "type": "string"
                }
            }
        },
...
        "dto.Coordinates": {
            "type": "object",
            "properties": {
                "x": {
                    "type": "string"
                },
                "y": {
                    "type": "string"
                }
            }
        },
        "dto.Count": {
            "type": "object",
            "required": [
                "count"
            ],
            "properties": {
                "count": {
                    "type": "string"
                }
            }
        },
        "dto.CountResponse": {
            "type": "object",
            "properties": {
                "count": {},
                "data": {}
            }
        },
        "dto.CourierDelivery": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "apartment_office": {
                    "type": "string"
                },
                "entrance": {
                    "type": "string"
                },
                "floor": {
                    "type": "string"
                },
                "index": {
                    "type": "string"
                },
                "intercom": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                }
            }
        },
...
        "dto.DefaultData": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
...
        "dto.DeliveryType": {
            "type": "object",
            "properties": {
                "cdek_delivery": {
                    "$ref": "#/definitions/dto.CDEKDelivery"
                },
                "courier_delivery": {
                    "$ref": "#/definitions/dto.CourierDelivery"
                },
                "self_delivery": {
                    "$ref": "#/definitions/dto.SelfDelivery"
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "dto.GetOrder": {
            "type": "object",
            "properties": {
                "canceled": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.OrderProduct"
                    }
                },
                "status": {
                    "type": "string"
                },
                "total_price": {
                    "type": "string"
                }
            }
        },
...
        "dto.Order": {
            "type": "object",
            "required": [
                "delivery_type",
                "payment_method_id"
            ],
            "properties": {
                "delivery_type": {
                    "$ref": "#/definitions/dto.DeliveryType"
                },
                "payment_method_id": {
                    "type": "integer"
                },
                "promo_code": {
                    "type": "string"
                }
            }
        },
        "dto.OrderProduct": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                },
                "quantity": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "total_price": {
                    "type": "string"
                }
            }
        },
...
        "dto.PaymentMethod": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "dto.PickUpPoint": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "coordinates": {
                    "$ref": "#/definitions/dto.Coordinates"
                },
                "email_1": {
                    "type": "string"
                },
                "email_2": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "phone_1": {
                    "type": "string"
                },
                "phone_2": {
                    "type": "string"
                },
                "phone_3": {
                    "type": "string"
                },
                "stock": {
                    "$ref": "#/definitions/dto.PickUpPointStockTitle"
                },
                "time": {
                    "$ref": "#/definitions/dto.PickUpPointTime"
                }
            }
        },
        "dto.PickUpPointStockDescription": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                }
            }
        },
        "dto.PickUpPointStockTitle": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PickUpPointStockDescription"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "dto.PickUpPointTime": {
            "type": "object",
            "properties": {
                "fri": {
                    "type": "string"
                },
                "mon": {
                    "type": "string"
                },
                "sat": {
                    "type": "string"
                },
                "sun": {
                    "type": "string"
                },
                "thu": {
                    "type": "string"
                },
                "tue": {
                    "type": "string"
                },
                "wen": {
                    "type": "string"
                }
            }
        },
...
        "dto.Receipt": {
            "type": "object",
            "properties": {
                "cart_price": {
                    "type": "string"
                },
                "delivery_price": {
                    "type": "string"
                },
                "final_price": {
                    "type": "string"
                },
                "products_count": {
                    "type": "string"
                },
                "sale": {
                    "type": "string"
                }
            }
        },
...
        "dto.ReturnOrder": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "dto.SelfDelivery": {
            "type": "object",
            "properties": {
                "pick_up_points_id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.UserOrdersProducts": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "quantity": {
                    "type": "string"
                },
                "rating": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "total_price": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}